Refactor(auth): application 계층 DTO 폴더명 변경
