+++ auth-service/src/main/java/com/rushcrew/auth_service/auth/application/client/UserClient.java
@@ -1,7 +1,7 @@
 package com.rushcrew.auth_service.auth.application.client;
 
-import com.rushcrew.auth_service.auth.application.dto.SignUpCommand;
-import com.rushcrew.auth_service.auth.application.dto.UserCreateResult;
+import com.rushcrew.auth_service.auth.application.command.SignUpCommand;
+import com.rushcrew.auth_service.auth.application.result.UserCreateResult;
 
 public interface UserClient {
     UserCreateResult createUser(SignUpCommand command);
