+++ auth-service/src/main/java/com/rushcrew/auth_service/auth/application/command/SignUpCommand.java
@@ -1,4 +1,4 @@
-package com.rushcrew.auth_service.auth.application.dto;
+package com.rushcrew.auth_service.auth.application.command;
 
 public record SignUpCommand(
     String email,
