+++ auth-service/src/main/java/com/rushcrew/auth_service/auth/application/result/UserCreateResult.java
@@ -1,4 +1,4 @@
-package com.rushcrew.auth_service.auth.application.dto;
+package com.rushcrew.auth_service.auth.application.result;
 
 public record UserCreateResult(
     Long userId,
