+++ auth-service/src/main/java/com/rushcrew/auth_service/auth/infrastructure/external/UserClientImpl.java
@@ -1,8 +1,8 @@
 package com.rushcrew.auth_service.auth.infrastructure.external;
 
 import com.rushcrew.auth_service.auth.application.client.UserClient;
-import com.rushcrew.auth_service.auth.application.dto.SignUpCommand;
-import com.rushcrew.auth_service.auth.application.dto.UserCreateResult;
+import com.rushcrew.auth_service.auth.application.command.SignUpCommand;
+import com.rushcrew.auth_service.auth.application.result.UserCreateResult;
 import com.rushcrew.auth_service.auth.infrastructure.external.dto.UserCreateRequest;
 import com.rushcrew.auth_service.auth.infrastructure.external.dto.UserCreateResponse;
 import lombok.RequiredArgsConstructor;
